data_collector/requirements.txt
grpcio
grpcio-tools
flask
flask-sqlalchemy
Flask-APScheduler
pymysql
requests